Salt subsystems self-contained - no longer require changes to salt/loader.py
Since the dawn of time I have been annoyed that new salt subsystems/extension types
required new entries to salt/loader.py This causes loader.py to be larger than necessary
as well as creates a nexus for revision contention.  Making salt.loader probe for
extension directories that self-provide LazyLoader invocations makes salt more modular.
* * *
Fix some _module_dirs => module_dirs
* * *
Avoid shadowing module_dirs symbol.
